+++ .gitmodules
@@ -1,3 +1,6 @@
 [submodule "2/libs/SafeInt"]
 	path = 2/libs/SafeInt
 	url = https://github.com/dcleblanc/SafeInt.git
+[submodule "2/libs/googletest"]
+	path = 2/libs/googletest
+	url = https://github.com/google/googletest.git
