+++ 2/CMakeLists.txt
@@ -1,14 +1,16 @@
 cmake_minimum_required(VERSION 3.17)
-project(calc CXX)
+project(Calc)
 
 set(CMAKE_CXX_STANDARD 17)
 
+add_subdirectory(src/calculator)
+add_subdirectory(src/commands)
+
+enable_testing()
 add_subdirectory(libs/SafeInt)
-add_executable(calc main.cpp src/Command.cpp src/Command.h src/Exceptions.cpp src/Exceptions.h)
+add_subdirectory(libs/googletest)
+add_subdirectory(tests)
 
-
-
-#target_link_libraries(calc SafeInt)
 if (MSVC)
     target_compile_options(calc
             PRIVATE /W4
