+++ 2/src/calculator/CMakeLists.txt
@@ -1,0 +1,5 @@
+add_executable(calc main.cpp)
+
+target_include_directories(calc PRIVATE "../commands")
+
+target_link_libraries(calc commands)
