+++ 2/src/commands/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_library(commands STATIC Command.h Command.cpp Exceptions.h Exceptions.cpp)
