+++ 2/src/commands/Command.cpp
@@ -36,6 +36,10 @@ void Context::push(int64_t val) {
     calc_stack_avokado.push(val);
 }
 
+Context::Context(ostream &output) : output(output) {}
+
+
+
 
 void BinaryArithmeticalCommand::eval() {
     SafeInt<int64_t> first = context.pop();
@@ -80,7 +84,7 @@ void Read::eval() {
 Read::Read(Context &curContext) : NonArgsCommand(curContext) {}
 
 void Print::eval() {
-    cout << context.top() << endl;
+    context.output << context.top() << endl;
 }
 
 Print::Print(Context &curContext) : NonArgsCommand(curContext) {}
@@ -102,3 +106,53 @@ void Push::eval(string &args) {
 }
 
 ArithmeticalCommand::ArithmeticalCommand(Context &curContext) : NonArgsCommand(curContext) {}
+
+shared_ptr<Command> get_command_by_string(string &cur_string, Context &context) {
+    shared_ptr<Command> operation;
+    if (cur_string.empty()) {
+        operation = nullptr;
+    } else if (cur_string == "POP") {
+        operation = shared_ptr<Command>(new Pop(context));
+    } else if (cur_string == "PUSH") {
+        operation = shared_ptr<Command>(new Push(context));
+    } else if (cur_string == "PEEK") {
+        operation = shared_ptr<Command>(new Peek(context));
+    } else if (cur_string == "ABS") {
+        operation = shared_ptr<Command>(new Abs(context));
+    } else if (cur_string == "PLUS") {
+        operation = shared_ptr<Command>(new Plus(context));
+    } else if (cur_string == "MINUS") {
+        operation = shared_ptr<Command>(new Minus(context));
+    } else if (cur_string == "MUL") {
+        operation = shared_ptr<Command>(new Mul(context));
+    } else if (cur_string == "DIV") {
+        operation = shared_ptr<Command>(new Div(context));
+    } else if (cur_string == "PRINT") {
+        operation = shared_ptr<Command>(new Print(context));
+    } else if (cur_string == "READ") {
+        operation = shared_ptr<Command>(new Read(context));
+    } else {
+        if (cur_string != "#") {
+            throw UnknownArgument("Unknown command: " + cur_string);
+        } else {
+            operation = shared_ptr<Command>(new Comment(context));
+        }
+    }
+    return operation;
+}
+
+
+void parse_stream(istream &stream, Context &context) {
+    while (!stream.eof()) {
+        string strCommand;
+        stream >> strCommand;
+        shared_ptr<Command> command = get_command_by_string(strCommand, context);
+        if (auto argsCommand = dynamic_pointer_cast<ArgsCommand>(command)) {
+            string args;
+            getline(stream, args);
+            argsCommand->eval(args);
+        } else if (auto nonArgsCommand = dynamic_pointer_cast<NonArgsCommand>(command)) {
+            nonArgsCommand->eval();
+        }
+    }
+}
