+++ 2/tests/CMakeLists.txt
@@ -1,0 +1,8 @@
+add_executable(calculator_tests tests.cpp)
+
+target_include_directories(calculator_tests PRIVATE "../src/commands")
+
+target_link_libraries(calculator_tests commands gtest gmock_main)
+
+enable_testing()
+add_test(NAME google_test COMMAND calculator_tests)
